Updated the site
Updated the entire site (I think, I've completely forgotten not going to lie). Fixed the broken VMware installer and added JewleahInstaller V3 into the downloads page and removed the beta of the same version, followed by keeping the old v2 download as a backup if something in v3 doesn't work but works fine in v2.